requirements.txt
streamlit==1.28.0
openai==0.28.0
gtts==2.4.0
PyPDF2==3.0.1
python-docx==1.0.1